@@ -25,11 +25,11 @@
         }
     },
     "require": {
-        "php": "^5.3.3|^7.0|^8.0",
-        "php-amqplib/php-amqplib": "^2.6"
+        "php": "^7.4|^8.0",
+        "php-amqplib/php-amqplib": "^3.5"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8",
+        "phpunit/phpunit": "^9.1",
         "scrutinizer/ocular": "^1.2",
         "squizlabs/php_codesniffer": "^2.6"
     },
