@@ -2,11 +2,13 @@
 
 namespace Thumper\Test;
 
-abstract class BaseTest extends \PHPUnit_Framework_TestCase
+use \PHPUnit\Framework\TestCase;
+
+abstract class BaseTest extends TestCase
 {
     /**
      * @param string[] $methods
-     * @return \PhpAmqpLib\Connection\AbstractConnection|\PHPUnit_Framework_MockObject_MockObject
+     * @return \PhpAmqpLib\Connection\AbstractConnection|\PHPUnit\Framework\MockObject\MockObject
      */
     protected function getMockConnection($methods = null)
     {
@@ -20,11 +22,11 @@ abstract class BaseTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @return \PhpAmqpLib\Channel\AMQPChannel|\PHPUnit_Framework_MockObject_MockObject
+     * @return \PhpAmqpLib\Channel\AMQPChannel|\PHPUnit\Framework\MockObject\MockObject
      */
     protected function getMockChannel()
     {
-        return $this->getMockBuilder('\PhpAmqpLib\Channel\AMQPChannel')
+        return $this->getMockBuilder(\PhpAmqpLib\Channel\AMQPChannel::class)
             ->disableOriginalConstructor()
             ->getMock();
     }
